Replace unitok and pymajka with pythonic versions
Previously, unitok was used via execve() what lead to another run of whole python. With
majka situation was a bit better because it was C-executable. This commit uses unitok as a module
and majka is used via C library bindings.

This changes required to have unicode string instead of 'str', so changes are included in this commit too.